# feat: run policy queries in parallel