--- a/sg/internal/engine/limiter.go
+++ b/sg/internal/engine/limiter.go
@@ -5,7 +5,7 @@ import "runtime"
 // channelLimiter is a channel based implementation of Limiter.
 type channelLimiter chan struct{}
 
-var _ Limiter = (*channelLimiter)(nil)
+var _ limiter = (*channelLimiter)(nil)
 
 func (l channelLimiter) acquire() func() {
 	l <- struct{}{}
@@ -16,6 +16,6 @@ func newLimiter(concurrency int) channelLimiter {
 	return make(chan struct{}, concurrency)
 }
 
-func newCPUBoundedLimiter() channelLimiter {
+func newLimiterFromMaxProcs() channelLimiter {
 	return newLimiter(runtime.GOMAXPROCS(0))
 }