--- a/sg/internal/engine/limiter_test.go
+++ b/sg/internal/engine/limiter_test.go
@@ -11,21 +11,23 @@ func TestLimiter(t *testing.T) {
 	t.Run("newLimiter with 1 concurrency", func(t *testing.T) {
 		limiter := newLimiter(1)
 
+		times := 10
 		wg := &sync.WaitGroup{}
-		p := make(chan int, 10)
-		for i := 0; i < len(p); i++ {
+		p := make(chan int, times)
+		for i := 0; i < times; i++ {
 			i := i
 			wg.Add(1)
+			doneLimit := limiter.acquire()
 			go func() {
 				defer wg.Done()
-				defer limiter.acquire()()
+				defer doneLimit()
 
 				p <- i
 			}()
 		}
 		wg.Wait()
 
-		for i := 0; i < len(p); i++ {
+		for i := 0; i < times; i++ {
 			assert.Equal(t, i, <-p)
 		}
 	})