--- a/sg/internal/engine/rego_query.go
+++ b/sg/internal/engine/rego_query.go
@@ -45,7 +45,7 @@ const PackageMain = "main"
 type RegoEngine struct {
 	policyPackages []policy.Package
 	compiler       *ast.Compiler
-	limiter        Limiter
+	limiter        limiter
 }
 
 var _ Queryer = (*RegoEngine)(nil)